requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.1
backcall==0.2.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
decorator==4.4.2
distlib==0.3.1
Django==3.1.5
django-mysql==3.11.1
django-redis==4.12.1
djangorestframework==3.12.2
filelock==3.0.12
identify==2.1.0
importlib-metadata==3.7.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
nodeenv==1.5.0
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pre-commit==2.10.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
Pygments==2.8.0
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
virtualenv==20.4.2
wcwidth==0.2.5
zipp==3.4.1
